examples/outbound_smb_offices/pyproject.toml
[project]
name = "basic_chat"
version = "0.1.0"
description = "Single system prompt voice agent that uses Gemini 2.5 Flash"
requires-python = ">=3.10"
dependencies = [
    "cartesia-line",
    "google-genai>=1.26.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "aiohttp>=3.12.0",
    "uvicorn==0.35.0",
    "fastapi>=0.104.0",
    "pydantic-partial==0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "pytest-cov==7.0.0",
    "pytest-xdist==3.8.0",
    "pytest-repeat==0.9.4"
]

# [tool.uv.sources]
# cartesia-line = { path = "../../", editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv.sources]
cartesia-line = { path = "/Users/nathan/workspace/line", editable = true }
